annoying relations fix

+++ Infrastructure/Repositories/UserRepo.cs
@@ -66,21 +66,32 @@ namespace AeldreplejeInfrastructure.Repositories
 
         public User UpdateUser(User user)
         {
-            var newPShiftList = new List<UserPendingShift>(user.PShifts);
-            _context.Attach(user).State = EntityState.Modified;
-            _context.Entry(user).Collection(u => u.Shifts).IsModified = true;
-            _context.Entry(user).Reference(u => u.Group).IsModified = true;
-
-            var shifts = _context.Shifts.Where(s => s.User.Id == user.Id && !user.Shifts.Exists(us => us.Id == s.Id));
-
-            foreach (var shift in shifts)
+            if(user.PShifts != null)
             {
-                shift.User = null;
-                _context.Entry(shift).Reference(s => s.User).IsModified = true;
+                var newPShiftList = new List<UserPendingShift>(user.PShifts);
             }
+            
+            _context.Attach(user).State = EntityState.Modified;
+            _context.Entry(user).Reference(u => u.Group).IsModified = true;
+            if(user.Shifts != null)
+            {
+                _context.Entry(user).Collection(u => u.Shifts).IsModified = true;
 
-            _context.UserPendingShifts.RemoveRange(
+                var shifts = _context.Shifts.Where(s => s.User.Id == user.Id && !user.Shifts.Exists(us => us.Id == s.Id));
+
+                foreach (var shift in shifts)
+                {
+                    shift.User = null;
+                    _context.Entry(shift).Reference(s => s.User).IsModified = true;
+                }
+            }
+            
+
+            if (user.PShifts != null)
+            {
+                _context.UserPendingShifts.RemoveRange(
                 _context.UserPendingShifts.Where(usp => usp.UserId == user.Id));
+            }
 
 
             _context.SaveChanges();
